feat: Add timer functionality, image picker, and payment container
This commit adds the timer functionality to the app using the `circular_countdown_timer` package and a new `TimerCubit`. It also adds the ability to pick images using the `image_picker` package and a new `pickImageFromGallery` method in the `AuthCubit`. Additionally, a new `PaymentContainer` widget has been added to display payment information. The `search_view` has been updated to display the `PaymentContainer` when a garage is tapped. Finally, some minor UI updates have been made to the `profile_view`.